src/requirements.txt
pdfminer.six
html2text
beautifulsoup4
PyYAML
jsonschema
pytest
